requirements.txt
requests
beautifulsoup4
apscheduler
python-dotenv
google-genai
pip-system-certs
